# daemon/src/routes/version.rs
use axum::Json;
use serde::Serialize;
use utoipa::ToSchema;

#[derive(Serialize, ToSchema)]
pub struct VersionResponse {
    version: &'static str,
}

#[utoipa::path(
    get,
    path = "/version",
    tag = "Version",
    responses((status = 200, description = "Service version", body = VersionResponse))
)]
pub async fn version() -> Json<VersionResponse> {
    Json(VersionResponse {
        version: env!("CARGO_PKG_VERSION"),
    })
}